[direct] rename kyo.directInternal to kyo.internal
