+++ kyo-direct/shared/src/main/scala/kyo/internal/Trees.scala
@@ -1,4 +1,4 @@
-package kyo.directInternal
+package kyo.internal
 
 import scala.quoted.*
